#3, #8: Integrate zwo_fixer into capture, also add linker path magic
- Reworked the Makefile a fair bit. It's still an ugly mess that needs
to go away and get replaced by something; but it's better than it was.
- Added an include path directive hack for Arch Linux's sake (shouldn't
break anything else). Basically we really ought to be using pkg-config
for that kind of stuff, but we aren't, because the Makefile remains a
merely-partially-mitigated dumpster fire.
- Replaced idiocy like '-lpthread' with '-pthread' etc.

- #3: The capture binary now links against libzwo_fixer.so!
- Also, the capture binary now links against **an explicit particular
version** of the ASICamera2 library, since zwo_fixer is hand-tuned with
specific offsets obtained from the particular library version it was
built to un-screw-up.

- #8: LD_LIBRARY_PATH should be a thing of the past. Now when the
linking process is done, I use some -rpath directives to ensure that
when the runtime linker is invoked at program runtime, it will look for
the ASICamera2 and zwo_fixer libraries in the same relative paths where
they were at build time.
- For extra belt-and-suspenders capabilities, I also put additional
-rpath directives in there to the **absolute** paths to the library
files, so that even if you run the binary from a different directory,
the runtime linker will **still** know where to find them.