capture: include <zwo_fixer.hpp> and call ZWOFixerInit at startup
It appears that if the capture program didn't actually directly use
anything from libzwo_fixer.so, then the linker would cleverly deduce
that it obviously didn't need to actually link to libzwo_fixer.so at
all. Even though we asked it to, and even though there are things that
we implicitly want to happen by virtue of linking to the library. So now
we call ZWOFixerInit at startup time to ensure that the linker doesn't
try any funny business like that.