Implement ARMv7 support (untested)
ARM Thumb mode is the most evil thing ever conceived from a reverse
engineering point of view. Not only do you (and your disassembler) never
have a goddamn clue when the code is in Thumb mode or not (leading to
confusion and wrongly-disassembled instructions); but then if you want
to hack around with this code and actually interface with it, good luck
with that, because probably half the stuff will be in regular ARM ISA
and the other half will be in Thumb and it's just a giant inconsistent
mess.

You have no idea how happy I am that ARM made the highly intelligent
decision to NOT support the Thumb instruction set(s) in AArch64.
Dropping that shit was the best decision they ever made.