#60: Fix Makefile issues arising from IA32/AMD64 arch assumption
This Makefile is frankly an abomination. It's still originally based on
the crusty ugly Makefile provided with the demo programs in the ASK SDK.

The necessity to manually adjustment of the PLATFORM variable in the
Makefile is idiotic and really should not need to be there, except maybe
in cases of cross-compilation.